+++ src/xrGame/movement_manager.h
@@ -225,8 +225,8 @@ protected:
 
 protected:
     Fvector path_position(const float& time_to_check);
-    Fvector path_position(const float& velocity, const Fvector& position, const float& time_delta,
-        u32& current_travel_point, float& dist, float& dist_to_target, Fvector& dir_to_target);
+    Fvector path_position(const float& velocity, const Fvector& position,
+ const float& time_delta, size_t & current_travel_point, float& dist, float& dist_to_target, Fvector& dir_to_target);
 
 protected:
     virtual CRestrictedObject* create_restricted_object();
@@ -240,7 +240,7 @@ private:
 
 public:
     virtual const float& prediction_speed() const;
-    Fvector predict_position(const float& time_delta, const Fvector& position, u32& current_travel_point,
+    Fvector predict_position(const float& time_delta, const Fvector& position, size_t & current_travel_point,
         const float& prediction_speed) const;
     Fvector predict_position(const float& time_delta) const;
     Fvector target_position() const;
